modified CopyModal and DataFilesToolbar and removed MakePublic button
