+++ client/src/components/DataFiles/DataFilesToolbar/DataFilesToolbar.js
@@ -64,14 +64,6 @@ const DataFilesToolbar = ({ scheme, api }) => {
     state => state.workbench.config.compressApp && modifiableUserData
   );
 
-  const showMakePublic = useSelector(
-    state =>
-      state.workbench &&
-      state.workbench.config.makePublic &&
-      api === 'tapis' &&
-      modifiableUserData
-  );
-
   const toggleRenameModal = () =>
     dispatch({
       type: 'DATA_FILES_TOGGLE_MODAL',
@@ -128,16 +120,6 @@ const DataFilesToolbar = ({ scheme, api }) => {
     });
   };
 
-  const toggleMakePublicModal = () => {
-    dispatch({
-      type: 'DATA_FILES_TOGGLE_MODAL',
-      payload: {
-        operation: 'makePublic',
-        props: { selectedFile: selectedFiles[0] }
-      }
-    });
-  };
-
   const download = () => {
     if (canDownload) {
       dispatch({
@@ -174,8 +156,6 @@ const DataFilesToolbar = ({ scheme, api }) => {
   const canTrash = getFilePermissions('trash', permissionParams);
   const canCompress = getFilePermissions('compress', permissionParams);
   const canExtract = getFilePermissions('extract', permissionParams);
-  const canMakePublic =
-    showMakePublic && getFilePermissions('public', permissionParams);
   return (
     <>
       <div id="data-files-toolbar-button-row">
@@ -231,14 +211,6 @@ const DataFilesToolbar = ({ scheme, api }) => {
             iconName="link"
             onClick={toggleLinkModal}
             disabled={!canDownload}
-          />
-        )}
-        {showMakePublic && (
-          <ToolbarButton
-            text="Make Public"
-            iconName="publications"
-            onClick={toggleMakePublicModal}
-            disabled={!canMakePublic}
           />
         )}
         {showTrash && (
